requirements.txt
torch>=1.7.0
torchvision>=0.8.0
clip @ git+https://github.com/openai/CLIP.git
opencv-python>=4.5.0
pillow>=8.0.0
numpy>=1.19.0
pandas>=1.2.0
tqdm>=4.50.0
scipy>=1.6.0
transformers>=4.0.0
matplotlib>=3.3.0 